spotify_transformation_load_function/requirements.txt
google-cloud-bigquery
google-cloud-storage
spotipy
pandas
functions_framework